9 fucking hours of coding I'm gonna be dreaming about css to night..

+++ app/api/[storeId]/billboards/[billboardId]/route.ts
@@ -66,8 +66,23 @@ export async function DELETE (
       return new NextResponse("Unauthenticated", { status: 401 });
     }
 
-    if (!params.storeId) {
+    if (!params.billboardId) {
       return new NextResponse("Billboard ID is required", { status: 400 });
+    }
+
+    if (!params.storeId) {
+      return new NextResponse("Store ID is required", { status: 400 });
+    }
+
+    const storeByUserId = await prismadb.store.findFirst({
+      where: {
+        id: params.storeId,
+        userId
+      }
+    });
+
+    if (!storeByUserId) {
+      return new NextResponse("Unauthorized", { status: 403 });
     }
 
     const billboard = await prismadb.billboard.deleteMany({
